Create global postgres_helper and import it at multiple locations
